Introduce role: ableton.clang
This role will be used in cidre in place of teebor_choka.clang, since
we have some "philosophical differences" in how that role is
constructed. :) Namely, the architecture of the role does not
correctly support several clang versions, and it is not idempotent
because of the way that it installs packages.